+++ package.js
@@ -8,6 +8,9 @@ Package.describe({
 
 Package.onUse( function( api ){
     configure( api );
+    api.export([
+        'PlusButton'
+    ]);
     api.mainModule( 'src/client/js/index.js', 'client' );
 });