+++ src/client/components/plusButton/plusButton.js
@@ -13,18 +13,20 @@ Template.plusButton.onCreated( function(){
     const self = this;
     self.PCK = {
         defaults: {
-            classes: 'btn btn-primary'
+            classes: 'btn btn-primary',
+            shape: PlusButton.C.Shape.CIRCLE
         }
     };
 });
 
 Template.plusButton.helpers({
     // set button classes
-    disabled(){
-        const classes = Template.instance().PCK.defaults.classes;
+    classes(){
+        let classes = Template.instance().PCK.defaults.classes;
         if( Object.keys( Template.currentData()).includes( 'classes' )){
             classes = Template.currentData().classes;
         }
+        console.debug( 'returning', classes );
         return classes;
     },
 
@@ -36,5 +38,27 @@ Template.plusButton.helpers({
             disabled = 'disabled';
         }
         return disabled;
+    },
+
+    // choose the shape of the button, providing an ad-hoc class
+    shape(){
+        let shape = Template.instance().PCK.defaults.shape;
+        if( Object.keys( Template.currentData()).includes( 'shape' )){
+            const value = Template.currentData().shape;
+            let found = false;
+            Object.keys( PlusButton.C.Shape ).every(( k ) => {
+                if( PlusButton.C.Shape[k] === value ){
+                    found = true;
+                    return false;
+                }
+                return true;
+            });
+            if( found ){
+                shape = value;
+            } else {
+                console.error( 'pwix:plus-button unknown shape \''+value+'\'' );
+            }
+        }
+        return shape;
     }
 });