+++ src/client/js/constants.js
@@ -1,0 +1,11 @@
+/*
+ * pwix:plus-button/src/client/js/constants.js
+ */
+
+PlusButton.C = {
+    // button shape
+    Shape: {
+        CIRCLE: 'CIRCLE',
+        RECTANGLE: 'RECTANGLE'
+    }
+};