+++ src/client/js/global.js
@@ -1,0 +1,5 @@
+/*
+ * pwix:plus-button/src/client/js/global.js
+ */
+
+PlusButton = {};