+++ src/client/js/index.js
@@ -1,3 +1,7 @@
 /*
  * pwix:plus-button/src/client/js/index.js
  */
+
+import './global.js';
+//
+import './constants.js';